etl_process/requirements.txt
numpy==2.3.4
pandas==2.3.3
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.3
schedule==1.2.2
six==1.17.0
tzdata==2025.2
